Debug

+++ 11-combo/combo.cpp
@@ -18,10 +18,10 @@ std::vector<string> permutations(string word) {
   std::vector<string> result;
 
   for(string perm : perms) {
-    for(int i = 0; i < perms.size() + 1; i++) {
+    for(int i = 0; i < (int)perms.size() + 1; i++) {
       std::cout << i << std::endl;
 
-      if( i <= perms.size()) {
+      if(i <= (int)perms.size()) {
         string front = perm.substr(0, i);
         string rest = perm.substr(i, perm.size());
         result.push_back(front + current_char + rest);
